Major modifications of RTL structure, and corresponding changes.
Circuit and Synchronous are almost entirely rewritten.

Comb support has been removed, but will be readded soon.
Latches have also been removed, but these will not be reimplemented
in the near future. The new model will allow extensions for
asynchronous circuitry to be added easily, so they do not need to be
explicitly included at this time.