Replaced Logic newtype with Readable typeclass.
Readable has one method read, which returns an Expr. This name conflicts
with Prelude.read, but for now I don't see a better alternative.
The class is defined with multiple parameters in order to support some
more complex types it is defined for later, which don't have Kind *->*

readWire and readReg (and rd) are replaced by Readable instances.

(comb is only removed here for structural reasons, it is redefined
later in the finished file)