Replaced Num instance with a new class Arith.
We want to be able to compose Readable values of different types, but
the signature of Num does not allow this. Additionally, we would like
better control over what types can be added, and how.

The default implementations assume the wires carry twos-complement
integers and can be added directly in Verilog.

An instance for Int is included.