Completely replaced Interface system.
Interfaces can now be used in either direction.
This unifies Input/Output into Signal (and Output replaces the Signal
type in Logic) and InChan/OutChan into Channel.
This will also allow Readable to be removed in future.

It may also be possible to define an operator to bind any interface
with its inverted version, which would provide a more general and
useful version of the current (=:) operator.

Currently this uses a slightly awkward Direction constraint, which
must be applied to any direction parameter in Interface definitions.
An alternative to this would be nice, but it is adequate for now.