Replaced Data.Bits with Mantle.Bits.
This now gives a representation for types which permits Verilog vectors
to be included. This also replaces the representation of values in the
Expr tree with a new Value datatype which uses an Integer representation,
since this is currently the most common way to create literals. A binary
option for this type may be added in future if it becomes useful.

repType gives a VType for a value, can safely be passed undefined.
repExpr takes a value and gives an Expr representing it.